Create collectionsofDataWithArray.js
Arrays are the ordered collections of the value.  Array is used to store the important information.to add the element last to the array we can do like [ arrayname.length]. push to add to end, pop to remove from end, shift remove from start , unshift add to start. Concat is used to merge arrays.Includes Look for a value and indexOf just like str.indexOf it tells about the value of the array.reverse is used to reverse an array and join is used to creates a string form array.